✅ Add testRevert_InvalidNonce
This test ensures that the number used once AKA unique number AKA nonce for  this transaction is correct. In this test, the owner has not completed any previous transactions, so the nonce is expected to be zero, except we set the nonce to 1 in the SigUtils.Permit({...}) struct, so the permit method reverts. Yay!